Update ParkingCoin.sol
This code file is written to use the official utility token for the Smart Parking System. It is a standard ERC20 token designed to handle all value transfers within the application, such as paying for parking fees. In this code we use to import the code form famous github repository with which we can easily create our own ERC-20 token contract, which will be used to track an internal currency in a hypothetical smart city.